Removing original test_proximity script for Moez updated one.
